requirements.txt
fastapi
openai
uvicorn
httpx
python-dotenv
redis
firebase-admin